ReadMe, dependencies.bat and BuildScript
